tests dir structure upd
